+++ app/lib/actions.ts
@@ -4,10 +4,10 @@ import { z } from "zod";
 import { sql } from "@vercel/postgres";
 import { redirect } from "next/navigation";
 import bcrypt from "bcryptjs";
-
 const passwordValidation = new RegExp(
   /(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).*/
 );
+import { Recipe } from "@/app/lib/definitions";
 
 const AuthenticationSchema = z.object({
     email: z.string().email({ message: "Invalid email address" }),
@@ -105,4 +105,23 @@ export async function getUser(email: string) {
     const user = await sql`SELECT * FROM users WHERE email = ${email}`;
     return user.rows[0];
 }
-  
+
+export async function getRecipes(): Promise<Recipe[]> {
+    try {
+        const result = await sql`SELECT * FROM recipe ORDER BY created_on DESC`;
+        return result.rows.map(row => ({
+            recipe_id: row.recipe_id,
+            recipe_name: row.recipe_name,
+            recipe_servings: row.recipe_servings,
+            recipe_description: row.recipe_description,
+            recipe_instructions: row.recipe_instructions,
+            meal_type: row.meal_type,
+            recipe_image: row.recipe_image || null,
+            created_on: row.created_on,
+            user_id: row.user_id,
+        })) as Recipe[];
+    } catch (error) {
+        console.error("Error fetching recipes:", error);
+        return [];
+    }
+}
